requirements.txt
aiohttp
shapely
azure-storage-blob
GDAL
python-dotenv
tqdm